Changed name->label, started on wifi mgmt packets
I'm not a fan of label over name, esp given the tags, labels, and "tag labels" confusion, but figured we should align with the "official" name to avoid confusion when Lifx Labs' SDK hits.

Starting to introduce a lot of struct duplicity... perhaps we need to look into wiki pages? Or a structures area?